backend/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
asgiref==3.4.0
    # via
    #   django
    #   djoser
certifi==2021.5.30
    # via requests
cffi==1.14.5
    # via cryptography
chardet==4.0.0
    # via requests
coreapi==2.3.3
    # via djoser
coreschema==0.0.4
    # via coreapi
cryptography==3.4.7
    # via social-auth-core
defusedxml==0.7.1
    # via
    #   python3-openid
    #   social-auth-core
django==3.2.4
    # via
    #   -r requirements.in
    #   djangorestframework
    #   djangorestframework-simplejwt
django-templated-mail==1.1.1
    # via djoser
djangorestframework==3.12.4
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==4.7.1
    # via djoser
djoser==2.1.0
    # via -r requirements.in
factory-boy==3.2.0
    # via -r requirements.in
faker==8.9.1
    # via factory-boy
idna==2.10
    # via requests
itypes==1.2.0
    # via coreapi
jinja2==3.0.1
    # via coreschema
markupsafe==2.0.1
    # via jinja2
oauthlib==3.1.1
    # via
    #   requests-oauthlib
    #   social-auth-core
pillow==8.3.0
    # via -r requirements.in
pycparser==2.20
    # via cffi
pyjwt==2.1.0
    # via
    #   djangorestframework-simplejwt
    #   social-auth-core
python-dateutil==2.8.1
    # via faker
python3-openid==3.2.0
    # via social-auth-core
pytz==2021.1
    # via django
requests==2.25.1
    # via
    #   coreapi
    #   requests-oauthlib
    #   social-auth-core
requests-oauthlib==1.3.0
    # via social-auth-core
six==1.16.0
    # via
    #   python-dateutil
    #   social-auth-app-django
social-auth-app-django==4.0.0
    # via djoser
social-auth-core==4.1.0
    # via social-auth-app-django
sqlparse==0.4.1
    # via django
text-unidecode==1.3
    # via faker
uritemplate==3.0.1
    # via coreapi
urllib3==1.26.6
    # via requests
